+++ src/components/users.vue
@@ -4,18 +4,18 @@
   <div class="p-4">
      <div class="grid grid-cols-3 my-4">
         <div class="">
-          <input type="text" class="w-full px-4 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500" v-model="search" placeholder="Search" autocomplete="off" required/>
+          <input type="text" class="w-full px-4 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-[#00416ada]" v-model="search" placeholder="Search" autocomplete="off" required/>
         </div>   
 
         <div class="col-start-3 text-right">            
-            <button class="btn-sm bg-blue-600 text-white" @click="createUser()"><span class="material-icons mi-sm">add</span> New User</button>
+            <button class="btn-sm bg-[#00416ada] text-white" @click="createUser()"><span class="material-icons mi-sm">add</span> New User</button>
         </div>
     </div>
 
     <!-- table -->
     <table class="table-striped table-bordered bg-white">
         <thead class="sticky -top-4 z-10 bg-white">
-            <tr class="text-blue-600">
+            <tr class="text-[#00416ada]">
               <th>#</th>
               <th class="cursor-pointer" @click="sortTable('username')">User Name<span class="material-icons mi-md ml-1">{{ getSortIcon('username') }}</span></th>
               <th class="cursor-pointer" @click="sortTable('designation')">Designation<span class="material-icons mi-md ml-1">{{ getSortIcon('designation') }}</span></th>
@@ -49,15 +49,15 @@
 
       <template #body>
         <label class="block mb-2 mt-4">Username</label>
-        <input v-model="newUser.username" type="text" class="w-full px-4 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500" required/>
+        <input v-model="newUser.username" type="text" class="w-full px-4 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-[#00416ada]" required/>
 
         <label class="block mb-2 mt-4">Password</label>
-        <input v-model="newUser.password" type="password" class="w-full px-4 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500" required/>
+        <input v-model="newUser.password" type="password" class="w-full px-4 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-[#00416ada]" required/>
 
         <!-- <label class="block mb-2 mt-4">Designation</label>
-        <input v-model="newUser.designation" type="text" class="w-full px-4 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500" required/> -->
+        <input v-model="newUser.designation" type="text" class="w-full px-4 py-2 border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-[#00416ada]" required/> -->
         
-        <select v-model="newUser.designation" class="w-full px-4 py-2 mt-4 cursor-pointer border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500" required>
+        <select v-model="newUser.designation" class="w-full px-4 py-2 mt-4 cursor-pointer border border-gray-300 rounded-lg focus:outline-none focus:ring-2 focus:ring-[#00416ada]" required>
           <option disabled value="">Select Designation</option>
           <option>MANAGER</option>
           <option>BILL PERSON</option>
@@ -66,8 +66,8 @@
       </template>
 
       <template #footer>
-        <button @click="closeModal" type="button" class="btn-sm bg-gray-300 text-gray-700 hover:bg-gray-400">Cancel</button>
-        <button type="submit" class="btn-sm bg-blue-600 text-white hover:bg-blue-700">Save</button>
+        <button @click="closeModal" type="button" class="btn-md bg-gray-300 text-gray-700 hover:bg-gray-400">Cancel</button>
+        <button type="submit" class="btn-md bg-[#00416ada] text-white">Save</button>
       </template>
     </Modal>
   </form>
@@ -78,6 +78,7 @@
 import navBar from './navBar.vue'
 import axios from "axios";
 import Modal from "./ModalComponent.vue";
+import Swal from 'sweetalert2';
 
 export default {
   name: "UsersPage",
@@ -204,12 +205,25 @@ export default {
       }
     },
     async deleteUser(item) {
+      const result = await Swal.fire({
+        title: `Are you sure you want to delete the user "${item.username}"?`,
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#d33',
+        cancelButtonColor: '#898989',
+        confirmButtonText: 'Delete',
+        cancelButtonText: 'Cancel',
+      });
+
+      if (result.isConfirmed) {
         try {
-            await axios.delete(`http://localhost:3000/users/${item.id}`)
-            this.getUsers()
+          await axios.delete(`http://localhost:3000/users/${item.id}`)
+          this.getUsers()
         } catch (err) {
-            console.error(err)
+          console.error(err)
         }
+      }
+        
     }
   },
 };
